Removed Docusaurus // Added First Ideas
Docusaurus was removed, because of new plans to build feuerstein.dev with a static stite generator. For this reason the files from docusaurus were removed and important .md files were saved. On the other hand, there was some work done in things of the structure concept of the website. This was explained in the readme.md file